Move car trial debugger to App router
