+++ apps/store/src/features/carDealership/DebuggerCarTrial/debuggerCarTrial.types.ts
@@ -1,0 +1,5 @@
+export enum Field {
+  SSN = 'ssn',
+  RegistrationNumber = 'registrationNumber',
+  Tier = 'vtrCoverageCode',
+}
